Set options for complex pinentry programs
At least on Linux, without passing the display, ttyname, and ttytype
options, the pinentry program errors out when it tries to actually
display the pinentry program.

The list of options was taken from sketching on the set of options
that `gpg` passes to the agent - it's possible that LC_TYPE and
LC_MESSGAES aren't necessary, but I wasn't able to test all of the
pinentry alternative implementations.